Refactor model module (#130)
* Fix nanodrr imports

* Refactor initialization to a submodule

* Move network, scheduler, and loss to a submodule

* Make submodule for data ops

* Make submodule for data ops

* Go full tensor

* Remove unused augmentation

* Fix loss mean
