+++ src/xvr/model/__init__.py
@@ -1,4 +1,4 @@
-from .sampler import get_random_pose
+from .datum import get_random_pose
 from .trainer import Trainer
 
 __all__ = ["Trainer", "get_random_pose"]
