+++ src/xvr/model/datum/sampler.py
@@ -1,6 +1,5 @@
 import torch
 from jaxtyping import Float
-
 from nanodrr.camera import make_rt_inv
 from nanodrr.data import Subject
