+++ src/xvr/model/modules/__init__.py
@@ -1,0 +1,10 @@
+from .loss import PoseRegressionLoss
+from .network import PoseRegressor
+from .scheduler import IdentitySchedule, WarmupCosineSchedule
+
+__all__ = [
+    "PoseRegressionLoss",
+    "PoseRegressor",
+    "IdentitySchedule",
+    "WarmupCosineSchedule",
+]
