+++ src/xvr/model/modules/loss.py
@@ -1,7 +1,6 @@
 from typing import NamedTuple
 
 import torch
-
 from nanodrr.metrics import DoubleGeodesicSE3, MultiscaleNormalizedCrossCorrelation2d
