+++ src/xvr/model/modules/network.py
@@ -1,6 +1,5 @@
 import timm
 import torch
-
 from nanodrr.geometry import Parameterization, convert
